refactor(parser): extract inline text handling
